+++ modules/default/currentweather/currentweather.js
@@ -24,6 +24,7 @@ Module.register("currentweather",{
 		useBeaufort: true,
 		lang: config.language,
 		showHumidity: false,
+		degreeLabel: false,
 
 		initialLoadDelay: 0, // 0 seconds delay
 		retryDelay: 2500,
@@ -182,9 +183,24 @@ Module.register("currentweather",{
 		weatherIcon.className = "wi weathericon " + this.weatherType;
 		large.appendChild(weatherIcon);
 
+		var degreeLabel = "";
+		if (this.config.degreeLabel) {
+			switch (this.config.units ) {
+			case "metric":
+				degreeLabel = "C";
+				break;
+			case "imperial":
+				degreeLabel = "F";
+				break;
+			case "default":
+				degreeLabel = "K";
+				break;
+			}
+		}
+
 		var temperature = document.createElement("span");
 		temperature.className = "bright";
-		temperature.innerHTML = " " + this.temperature + "&deg;";
+		temperature.innerHTML = " " + this.temperature + "&deg;" + degreeLabel;
 		large.appendChild(temperature);
 
 		wrapper.appendChild(large);