+++ CHANGELOG.md
@@ -48,7 +48,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add  test for compliments module for parts of day
 - Korean Translation.
 - Added console warning on startup when deprecated config options are used
-
+- Add option to display temperature unit label to the current weather module
 
 ### Fixed
 - Update .gitignore to not ignore default modules folder.